+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0] - 2021-07-02
+
 ### Added
 
 - Reviews now show if they are from a similar productor if they were originally posted on a third party site